renaming, bug fix on risk sort order, fix spacing issues
