+++ src/components/PatientView/PatientViewBanner/PatientViewBanner.js
@@ -5,21 +5,21 @@ import patientProps from '../../../prop-types/patient';
 import PatientViewBannerSummary from './PatientViewBannerSummary';
 
 const PatientViewBanner = ({ patient }) => {
-	if (patient) {
-		return (
+  if (patient) {
+    return (
 			<div>
 				<PatientViewBannerSummary patient={patient} />
 			</div>
 		);
-	}
-	return (
+  }
+  return (
 		<div>
 		</div>
 	);
 };
 
 PatientViewBanner.propTypes = {
-	patient: patientProps.isRequired
+  patient: patientProps.isRequired
 };
 
 export default PatientViewBanner;
