+++ src/middlewares/restructure_patients.js
