+++ src/middlewares/restructure_risk_assessments.js
