+++ src/store/configureStore.js
@@ -2,20 +2,18 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import promiseMiddleware from 'redux-promise-middleware';
 
 import requestBundlesMiddleware from '../middlewares/request_bundles';
-import requestHuddlesMiddleware from '../middlewares/request_huddles';
-import requestPatientsMiddleware from '../middlewares/request_patients';
-// import requestPopulationsMiddleware from '../middlewares/request_populations';
-import requestRiskAssessmentsMiddleware from '../middlewares/request_risk_assessments';
+import restructureGroupsMiddleware from '../middlewares/restructure_groups';
+import restructurePatientsMiddleware from '../middlewares/restructure_patients';
+import restructureRiskAssessmentsMiddleware from '../middlewares/restructure_risk_assessments';
 import rootReducer from '../reducers';
 
 export default function configureStore(initialState) {
   let middleware = applyMiddleware(
     promiseMiddleware(),
     requestBundlesMiddleware,
-    requestHuddlesMiddleware,
-    requestPatientsMiddleware,
-    // requestPopulationsMiddleware,
-    requestRiskAssessmentsMiddleware
+    restructureGroupsMiddleware,
+    restructurePatientsMiddleware,
+    restructureRiskAssessmentsMiddleware
   );
 
   const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
